requirements.txt
torch==2.2.2
torchvision==0.17.2
fastapi==0.111.0
uvicorn==0.29.0
pillow==10.3.0
scikit-learn==1.4.2
mlflow==2.12.2
dvc==3.50.1
pytest==8.1.1
httpx==0.27.0
requests==2.31.0
prometheus-client==0.20.0
python-multipart==0.0.9
numpy==1.26.4
